destringify_tests
